Mean First Passage Time for markovchain

Given a markovchain object, this function calculates the expected steps to go from state i to j

meanFirstPassageTime(markovchain,destination_set)

Arguments

markovchain
the markovchain-class object

destination_set
the set of destination states or NULL (all states)

Value

a vector (if destination_set given) or a matrix (otherwise) of mean passage times

Details

if destination_set is one or more states, the mean first passage time from each remaining state to the given set is computed. If NULL, the full MFPT matrix is computed (with a different algorithm).

References

C. M. Grinstead and J. L. Snell. Introduction to Probability. American Mathematical Soc., 2012.

Examples

Pmat <- matrix( c(6,3,1,  2,3,5, 4,1,5)/10, ncol=3, byrow=TRUE)
P <- new("markovchain", states=c("s","c","r"), transitionMatrix=Pmat)
meanFirstPassageTime(P,"r")

#>        s        c 
#> 4.545455 2.727273 

meanFirstPassageTime(P)

#>          s    c        r
#> s 0.000000 3.75 4.545455
#> c 3.333333 0.00 2.727273
#> r 2.666667 5.00 0.000000

# Grinstead and Snell's "Oz weather" worked out example
Poz <- new("markovchain", states=c("s","c","r"),
           transitionMatrix=matrix(c(2,1,1, 2,0,2, 1,1,2)/4, byrow=TRUE, ncol=3))
meanFirstPassageTime(Poz)

#>          s c        r
#> s 0.000000 4 3.333333
#> c 2.666667 0 2.666667
#> r 3.333333 4 0.000000