Function to generate a sequence of states from homogeneous or non-homogeneous Markov chains.

Provided any markovchain or markovchainList objects, it returns a sequence of states coming from the underlying stationary distribution.

rmarkovchain(n, object, what = "data.frame", useRCpp = TRUE,
  parallel = FALSE, num.cores = NULL, ...)

Arguments

n
Sample size

object
Either a markovchain or a markovchainList object

what
It specifies whether either a data.frame or a matrix (each rows represent a simulation) or a list is returned.

useRCpp
Boolean. Should RCpp fast implementation being used? Default is yes.

parallel
Boolean. Should parallel implementation being used? Default is yes.

num.cores
Number of Cores to be used

...
additional parameters passed to the internal sampler

Value

Character Vector, data.frame, list or matrix

Details

When a homogeneous process is assumed (markovchain object) a sequence is sampled of size n. When a non - homogeneous process is assumed, n samples are taken but the process is assumed to last from the begin to the end of the non-homogeneous markov process.

Note

Check the type of input

References

A First Course in Probability (8th Edition), Sheldon Ross, Prentice Hall 2010

See also

Examples

# define the markovchain object
statesNames <- c("a", "b", "c")
mcB <- new("markovchain", states = statesNames,
   transitionMatrix = matrix(c(0.2, 0.5, 0.3, 0, 0.2, 0.8, 0.1, 0.8, 0.1),
   nrow = 3, byrow = TRUE, dimnames = list(statesNames, statesNames)))

# show the sequence
outs <- rmarkovchain(n = 100, object = mcB, what = "list")

#define markovchainList object
statesNames <- c("a", "b", "c")
mcA <- new("markovchain", states = statesNames, transitionMatrix =
   matrix(c(0.2, 0.5, 0.3, 0, 0.2, 0.8, 0.1, 0.8, 0.1), nrow = 3,
   byrow = TRUE, dimnames = list(statesNames, statesNames)))
mcB <- new("markovchain", states = statesNames, transitionMatrix =
   matrix(c(0.2, 0.5, 0.3, 0, 0.2, 0.8, 0.1, 0.8, 0.1), nrow = 3,
   byrow = TRUE, dimnames = list(statesNames, statesNames)))
mcC <- new("markovchain", states = statesNames, transitionMatrix =
   matrix(c(0.2, 0.5, 0.3, 0, 0.2, 0.8, 0.1, 0.8, 0.1), nrow = 3,
   byrow = TRUE, dimnames = list(statesNames, statesNames)))
mclist <- new("markovchainList", markovchains = list(mcA, mcB, mcC))

# show the list of sequence
rmarkovchain(100, mclist, "list")

#> [[1]]
#> [1] "b" "c" "b"
#> 
#> [[2]]
#> [1] "b" "c" "b"
#> 
#> [[3]]
#> [1] "b" "c" "b"
#> 
#> [[4]]
#> [1] "b" "c" "b"
#> 
#> [[5]]
#> [1] "b" "c" "b"
#> 
#> [[6]]
#> [1] "b" "c" "a"
#> 
#> [[7]]
#> [1] "c" "b" "b"
#> 
#> [[8]]
#> [1] "b" "c" "b"
#> 
#> [[9]]
#> [1] "b" "c" "c"
#> 
#> [[10]]
#> [1] "c" "b" "c"
#> 
#> [[11]]
#> [1] "b" "c" "c"
#> 
#> [[12]]
#> [1] "b" "b" "b"
#> 
#> [[13]]
#> [1] "b" "c" "c"
#> 
#> [[14]]
#> [1] "a" "c" "a"
#> 
#> [[15]]
#> [1] "b" "b" "c"
#> 
#> [[16]]
#> [1] "b" "b" "c"
#> 
#> [[17]]
#> [1] "c" "b" "c"
#> 
#> [[18]]
#> [1] "c" "b" "b"
#> 
#> [[19]]
#> [1] "b" "b" "c"
#> 
#> [[20]]
#> [1] "b" "c" "c"
#> 
#> [[21]]
#> [1] "b" "c" "b"
#> 
#> [[22]]
#> [1] "b" "c" "b"
#> 
#> [[23]]
#> [1] "b" "c" "b"
#> 
#> [[24]]
#> [1] "c" "b" "b"
#> 
#> [[25]]
#> [1] "b" "c" "b"
#> 
#> [[26]]
#> [1] "c" "b" "c"
#> 
#> [[27]]
#> [1] "b" "c" "b"
#> 
#> [[28]]
#> [1] "c" "c" "b"
#> 
#> [[29]]
#> [1] "b" "b" "b"
#> 
#> [[30]]
#> [1] "b" "c" "b"
#> 
#> [[31]]
#> [1] "c" "b" "c"
#> 
#> [[32]]
#> [1] "c" "b" "b"
#> 
#> [[33]]
#> [1] "c" "b" "c"
#> 
#> [[34]]
#> [1] "c" "b" "c"
#> 
#> [[35]]
#> [1] "c" "b" "c"
#> 
#> [[36]]
#> [1] "b" "c" "b"
#> 
#> [[37]]
#> [1] "b" "c" "b"
#> 
#> [[38]]
#> [1] "c" "b" "c"
#> 
#> [[39]]
#> [1] "a" "a" "b"
#> 
#> [[40]]
#> [1] "c" "c" "c"
#> 
#> [[41]]
#> [1] "a" "b" "b"
#> 
#> [[42]]
#> [1] "c" "b" "b"
#> 
#> [[43]]
#> [1] "b" "c" "b"
#> 
#> [[44]]
#> [1] "b" "c" "b"
#> 
#> [[45]]
#> [1] "b" "c" "b"
#> 
#> [[46]]
#> [1] "b" "c" "b"
#> 
#> [[47]]
#> [1] "a" "a" "b"
#> 
#> [[48]]
#> [1] "c" "b" "c"
#> 
#> [[49]]
#> [1] "b" "c" "c"
#> 
#> [[50]]
#> [1] "c" "a" "b"
#> 
#> [[51]]
#> [1] "b" "c" "b"
#> 
#> [[52]]
#> [1] "c" "b" "c"
#> 
#> [[53]]
#> [1] "c" "b" "c"
#> 
#> [[54]]
#> [1] "b" "c" "b"
#> 
#> [[55]]
#> [1] "b" "c" "b"
#> 
#> [[56]]
#> [1] "b" "b" "c"
#> 
#> [[57]]
#> [1] "c" "b" "b"
#> 
#> [[58]]
#> [1] "b" "c" "b"
#> 
#> [[59]]
#> [1] "c" "b" "c"
#> 
#> [[60]]
#> [1] "b" "c" "b"
#> 
#> [[61]]
#> [1] "b" "c" "b"
#> 
#> [[62]]
#> [1] "b" "c" "a"
#> 
#> [[63]]
#> [1] "b" "c" "b"
#> 
#> [[64]]
#> [1] "c" "b" "c"
#> 
#> [[65]]
#> [1] "b" "c" "b"
#> 
#> [[66]]
#> [1] "c" "b" "c"
#> 
#> [[67]]
#> [1] "b" "b" "c"
#> 
#> [[68]]
#> [1] "b" "c" "a"
#> 
#> [[69]]
#> [1] "c" "b" "b"
#> 
#> [[70]]
#> [1] "c" "b" "c"
#> 
#> [[71]]
#> [1] "a" "a" "a"
#> 
#> [[72]]
#> [1] "c" "a" "b"
#> 
#> [[73]]
#> [1] "b" "b" "b"
#> 
#> [[74]]
#> [1] "a" "c" "b"
#> 
#> [[75]]
#> [1] "c" "b" "c"
#> 
#> [[76]]
#> [1] "c" "b" "c"
#> 
#> [[77]]
#> [1] "c" "b" "c"
#> 
#> [[78]]
#> [1] "a" "c" "b"
#> 
#> [[79]]
#> [1] "c" "b" "c"
#> 
#> [[80]]
#> [1] "c" "b" "c"
#> 
#> [[81]]
#> [1] "a" "c" "b"
#> 
#> [[82]]
#> [1] "a" "b" "c"
#> 
#> [[83]]
#> [1] "b" "c" "c"
#> 
#> [[84]]
#> [1] "c" "b" "c"
#> 
#> [[85]]
#> [1] "a" "a" "b"
#> 
#> [[86]]
#> [1] "a" "c" "b"
#> 
#> [[87]]
#> [1] "b" "c" "a"
#> 
#> [[88]]
#> [1] "b" "c" "b"
#> 
#> [[89]]
#> [1] "c" "b" "b"
#> 
#> [[90]]
#> [1] "b" "c" "c"
#> 
#> [[91]]
#> [1] "c" "b" "c"
#> 
#> [[92]]
#> [1] "c" "b" "b"
#> 
#> [[93]]
#> [1] "b" "c" "b"
#> 
#> [[94]]
#> [1] "c" "a" "c"
#> 
#> [[95]]
#> [1] "b" "c" "b"
#> 
#> [[96]]
#> [1] "b" "c" "c"
#> 
#> [[97]]
#> [1] "a" "c" "c"
#> 
#> [[98]]
#> [1] "c" "b" "b"
#> 
#> [[99]]
#> [1] "b" "b" "c"
#> 
#> [[100]]
#> [1] "b" "c" "a"
#>